[python] Define buildModelsFromUrdf
